LayoutTests/fast/images/text-recognition/image-overlay-line-wrapping-expected.txt


PASS textarea.value is "this is a test\nfoo"
PASS successfullyParsed is true

TEST COMPLETE
